refactor(core): add root index file, consolidate utility imports and update paths
- Moved utility imports to a centralized index file in the utils directory.
- Updated import paths across various modules to use the new index file.
- Removed redundant utility functions and adjusted related code accordingly.
