+++ src/core/Field.ts
@@ -1,6 +1,6 @@
 import type { Resource } from "./Resource.js";
 import type { Nullable } from "./types.js";
-import { assignSealed } from "./utils/assignSealed.js";
+import { assignSealed } from "./utils/index.js";
 
 export type FieldType =
   | "string"
