+++ src/core/index.ts
@@ -1,7 +1,3 @@
-export * from "../graphql/index.js";
-export * from "../hydra/index.js";
-export * from "../openapi3/index.js";
-export * from "../swagger/index.js";
 export * from "./Api.js";
 export * from "./Field.js";
 export * from "./Operation.js";
