+++ src/core/utils/index.ts
@@ -1,0 +1,5 @@
+export { assignSealed } from "./assignSealed.js";
+export { getResourcePaths } from "./getResourcePaths.js";
+export { getType } from "./getType.js";
+export { parsedJsonReplacer } from "./parsedJsonReplacer.js";
+export { removeTrailingSlash } from "./removeTrailingSlash.js";
