+++ src/hydra/getParameters.ts
@@ -1,7 +1,7 @@
-import { Parameter } from "../Parameter.js";
+import type { Resource } from "../core/index.js";
+import { Parameter } from "../core/index.js";
+import type { RequestInitExtended } from "../core/types.js";
 import fetchResource from "./fetchResource.js";
-import type { Resource } from "../Resource.js";
-import type { RequestInitExtended } from "../types.js";
 
 export default async function getParameters(
   resource: Resource,
