+++ src/openapi3/handleJson.test.ts
@@ -1,7 +1,7 @@
-import handleJson from "./handleJson.js";
-import parsedJsonReplacer from "../utils/parsedJsonReplacer.js";
 import type { OpenAPIV3 } from "openapi-types";
 import { expect, test } from "vitest";
+import { parsedJsonReplacer } from "../core/utils/index.js";
+import handleJson from "./handleJson.js";
 
 const openApi3Definition: OpenAPIV3.Document = {
   openapi: "3.0.2",
